:sparkles: [Feat]: Entities 레이어에 schedule 슬라이스 추가
- 일정 관련 타입 및 api 코드
