+++ src/entities/schedule/api/api.ts
@@ -1,0 +1,16 @@
+export const scheduleApi = {
+  getTodaySchedules: async () => {},
+  getUpcomingSchedules: async () => {},
+  getSchedulesByMonth: async () => {},
+  getSchedulesByPeriod: async () => {},
+  getSchedulesByDateAndCalendar: async () => {},
+  getSchedulesByPeriodAndCalendar: async () => {},
+  getScheduleById: async () => {},
+
+  createSchedule: async () => {},
+  updateSchedule: async () => {},
+  deleteSchedule: async () => {},
+  shareSchedule: async () => {},
+
+  searchSchedules: async () => {},
+};
