+++ src/entities/schedule/api/index.ts
@@ -1,0 +1,2 @@
+export * from "./api";
+export * from "./types";
