+++ src/entities/schedule/api/types.ts
@@ -1,0 +1,3 @@
+import type { Schedule } from "../model";
+
+export type CreateScheduleRequest = Omit<Schedule, "id" | "calendar">;
