+++ src/entities/schedule/index.ts
@@ -1,0 +1,2 @@
+export { ScheduleApi } from "./api";
+export * from "./model";
