+++ src/entities/schedule/model/index.ts
